fail gracefully if no editors file

+++ .scripts/get_editors.py
@@ -16,30 +16,16 @@ if args.sourcedir == "no_dir_specified":
     print ("no_dir_specified")
     sys.exit()
 #-----------------------------
-
 edin = os.path.join(args.sourcedir,"offline_EDITORS.txt")
 edout = os.path.join(args.sourcedir,"editors.md")
-with open(edin) as f:
-    lines=f.readlines()
-now = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
-brackets = re.compile(r"\([^()]*\)$", re.IGNORECASE)
-with open(edout,"w") as o:
-    o.write("---\ntitle: Editors\n---\n\n# Editors\n\n")
-    o.write("Last updated {}\n\n".format(now))
-    for line in lines:
-        o.write("- {}".format(re.sub(brackets, "", line)))
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+if os.path.exists(edin):
+    with open(edin) as f:
+        lines=f.readlines()
+    now = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    brackets = re.compile(r"\([^()]*\)$", re.IGNORECASE)
+    with open(edout,"w") as o:
+        o.write("---\ntitle: Editors\n---\n\n# Editors\n\n")
+        o.write("Last updated {}\n\n".format(now))
+        for line in lines:
+            o.write("- {}".format(re.sub(brackets, "", line)))
